requirement.txt
langchain~=0.3.7
python-dotenv~=1.0.0
streamlit~=1.37.0
faiss-cpu==1.7.4
langchain-ollama==0.2.0
langchain-community==0.3.5
ffmpeg-python==0.2.0
ffmpeg==1.4
openai-whisper==20240930
